Add bookList array & add and remove reducer
