+++ src/bookList.js
@@ -1,0 +1,24 @@
+const bookList = [
+  {
+    item_id: 'item1',
+    title: 'The Magician',
+    author: 'Alain Quote',
+    category: 'Roman',
+  },
+
+  {
+    item_id: 'item2',
+    title: 'Second War',
+    author: 'Aime Loqw',
+    category: 'History',
+  },
+
+  {
+    item_id: 'item3',
+    title: 'Network Principile',
+    author: 'Josue Rushanika',
+    category: 'Technology',
+  },
+];
+
+export default bookList;
